move HIGGSDataLoader
