--- examples/advanced/random_forest/jobs/random_forest_base/app/config/config_fed_client.json
@@ -1,43 +1,0 @@
-{
-  "format_version": 2,
-
-  "executors": [
-    {
-      "tasks": [
-        "train"
-      ],
-      "executor": {
-        "id": "Executor",
-        "path": "nvflare.app_opt.xgboost.tree_based.executor.FedXGBTreeExecutor",
-        "args": {
-          "data_loader_id": "dataloader",
-          "training_mode": "bagging",
-          "num_client_bagging": 5,
-          "num_local_parallel_tree": 5,
-          "local_subsample": 0.8,
-          "lr_mode": "scaled",
-          "local_model_path": "model.json",
-          "global_model_path": "model_global.json",
-          "learning_rate": 0.1,
-          "objective": "binary:logistic",
-          "max_depth": 8,
-          "eval_metric": "auc",
-          "tree_method": "hist",
-          "use_gpus": false,
-          "nthread": 16
-        }
-      }
-    }
-  ],
-  "task_result_filters": [],
-  "task_data_filters": [],
-  "components": [
-    {
-      "id": "dataloader",
-      "path": "higgs_data_loader.HIGGSDataLoader",
-      "args": {
-        "data_split_filename": "data_split.json"
-      }
-    }
-  ]
-}
