--- examples/advanced/random_forest/jobs/random_forest_base/app/config/config_fed_server.json
@@ -1,52 +1,0 @@
-{
-  "format_version": 2,
-
-  "server": {
-    "heart_beat_timeout": 600,
-    "task_request_interval": 0.05
-  },
-
-  "task_data_filters": [],
-  "task_result_filters": [],
-
-  "components": [
-    {
-      "id": "persistor",
-      "path": "nvflare.app_opt.xgboost.tree_based.model_persistor.XGBModelPersistor",
-      "args": {
-        "save_name": "xgboost_model.json"
-      }
-    },
-    {
-      "id": "shareable_generator",
-      "path": "nvflare.app_opt.xgboost.tree_based.shareable_generator.XGBModelShareableGenerator",
-      "args": {}
-    },
-    {
-      "id": "aggregator",
-      "path": "nvflare.app_opt.xgboost.tree_based.bagging_aggregator.XGBBaggingAggregator",
-      "args": {}
-    }
-  ],
-  "workflows": [
-    {
-      "id": "scatter_and_gather",
-      "path": "nvflare.app_common.workflows.scatter_and_gather.ScatterAndGather",
-      "args": {
-        "min_clients": 5,
-        "num_rounds": 1,
-        "start_round": 0,
-        "wait_time_after_min_received": 0,
-        "aggregator_id": "aggregator",
-        "persistor_id": "persistor",
-        "shareable_generator_id": "shareable_generator",
-        "train_task_name": "train",
-        "train_timeout": 0,
-        "allow_empty_global_weights": true,
-        "task_check_period": 0.01,
-        "persist_every_n_rounds": 0,
-        "snapshot_every_n_rounds": 0
-      }
-    }
-  ]
-}
