--- examples/advanced/random_forest/jobs/random_forest_base/meta.json
@@ -1,9 +1,0 @@
-{
-  "name": "xgboost_tree_bagging",
-  "resource_spec": {},
-  "deploy_map": {
-    "app": [
-      "@ALL"
-    ]
-  }
-}
